Add drag-and-drop feature
Cards can be dragged from one CardSpot to another.

Currently, dragging a Card onto another Card will delete one of the Cards, including when a Card is dragged onto itself.

Spawning random Cards has been moved to double-click to avoid conflict with dragging.

Components are now in the function style in order to accommodate React Hooks, which are necessary for react-dnd.